Added call to `becomeFirstResponder` back in.
When using a swipe gesture, it is not guaranteed that the number input text field will become first responder unless this call is made. In order to avoid animation conflicts, the call to `becomeFirstResponder` is delayed until after the animation has completed (on a LTR script) or halfway through the animation (on a RTL script to provide a more responsive UI).